requirements.txt
Flask>=0.10.1
Flask-Login>=0.2.6
Flask-SQLAlchemy>=0.16
Flask-SeaSurf>=0.1.20
Flask-WTF>=0.8.4
Flask-Uploads>=0.1.3
Flask-Sijax>=0.3.1
Ghost.py>=0.1.b2
Jinja2>=2.7
MarkupSafe>=0.18
SQLAlchemy>=0.8.2
Sijax>=0.2.6
WTForms>=1.0.4
Werkzeug>=0.9.1
itsdangerous>=0.22
mysql-python>=1.2.4
